pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbotsclient"
version = "1.0.6"
dependencies = [
    "pusher", 
    "pysher", 
    "spacy", 
    "Flask",
    "Flask-Cors", 
    "porter2stemmer",
    "ccpke",
    "nltk"
]